# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Laporan Harian - Daily Work Report System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Department {
  id: number;
  name: string;
  code: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  departmentName?: string | null;
  avatarInitials: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface DailyTask {
  id: number;
  reportId: number;
  title: string;
  /** @nullable */
  project?: string | null;
  progress: number;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface TaskInput {
  title: string;
  project?: string;
  progress: number;
  status: string;
  notes?: string;
}

export interface TaskUpdate {
  title?: string;
  project?: string;
  progress?: number;
  status?: string;
  notes?: string;
}

export interface ReportComment {
  id: number;
  reportId: number;
  userId: number;
  userName: string;
  userRole: string;
  comment: string;
  createdAt: string;
}

export interface CommentInput {
  comment: string;
}

export interface Report {
  id: number;
  userId: number;
  userName: string;
  userEmail: string;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  departmentName?: string | null;
  date: string;
  dayName: string;
  /** @nullable */
  obstacles?: string | null;
  /** @nullable */
  additionalNotes?: string | null;
  /** @nullable */
  tomorrowPlan?: string | null;
  status: string;
  tasks: DailyTask[];
  comments: ReportComment[];
  taskCount: number;
  avgProgress: number;
  createdAt: string;
  updatedAt: string;
}

export interface ReportSummary {
  id: number;
  userId: number;
  userName: string;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  departmentName?: string | null;
  date: string;
  dayName: string;
  taskCount: number;
  avgProgress: number;
  status: string;
  createdAt: string;
}

export interface ReportInput {
  date: string;
  obstacles?: string;
  additionalNotes?: string;
  tomorrowPlan?: string;
  status?: string;
}

export interface ReportUpdate {
  date?: string;
  obstacles?: string;
  additionalNotes?: string;
  tomorrowPlan?: string;
  status?: string;
}

export interface ReviewInput {
  action: string;
  comment?: string;
}

export interface Notification {
  id: number;
  userId: number;
  title: string;
  message: string;
  isRead: boolean;
  type: string;
  /** @nullable */
  relatedReportId?: number | null;
  createdAt: string;
}

export interface DashboardSummary {
  totalEmployees: number;
  submittedToday: number;
  notSubmittedToday: number;
  totalTasksToday: number;
  tasksCompleted: number;
  tasksPending: number;
  submitRate: number;
  completionRate: number;
}

export interface DepartmentProductivity {
  departmentId: number;
  departmentName: string;
  employeeCount: number;
  submittedCount: number;
  avgProgress: number;
  submitRate: number;
}

export interface PoItem {
  id: number;
  noPo: string;
  namaProject: string;
  /** @nullable */
  customer?: string | null;
  tanggalPoMasuk?: string;
  /** @nullable */
  targetPenyelesaian?: string | null;
  deadline: string;
  sisaHari: number;
  /** @nullable */
  picUserId?: number | null;
  /** @nullable */
  picName?: string | null;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  departmentName?: string | null;
  status: string;
  progress: number;
  /** @nullable */
  catatan?: string | null;
  /** @nullable */
  createdByUserId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface PoInput {
  noPo: string;
  namaProject: string;
  customer?: string;
  tanggalPoMasuk: string;
  targetPenyelesaian?: string;
  deadline: string;
  picUserId?: number;
  departmentId?: number;
  status?: string;
  progress?: number;
  catatan?: string;
}

export interface PoUpdate {
  noPo?: string;
  namaProject?: string;
  customer?: string;
  tanggalPoMasuk?: string;
  targetPenyelesaian?: string;
  deadline?: string;
  picUserId?: number;
  departmentId?: number;
  status?: string;
  progress?: number;
  catatan?: string;
}

export interface PoSummary {
  totalPo: number;
  poSelesai: number;
  poBelumSelesai: number;
  poDelay: number;
  poHampirDeadline: number;
  persentasePencapaian: number;
  month: number;
  year: number;
}

export type ListReportsParams = {
date?: string;
month?: string;
year?: string;
/**
 * @nullable
 */
departmentId?: number | null;
/**
 * @nullable
 */
userId?: number | null;
status?: string;
search?: string;
};

export type ListPoParams = {
month?: number;
year?: number;
status?: string;
departmentId?: number;
picUserId?: number;
customer?: string;
search?: string;
};

export type GetPoSummaryParams = {
month?: number;
year?: number;
};

export type GetDashboardSummaryParams = {
date?: string;
};

export type GetDepartmentProductivityParams = {
date?: string;
};