# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Laporan Harian - Daily Work Report System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/{id}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a report
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReportComment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a report
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportComment"
...
  /notifications/{notifId}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: notifId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      parameters:
        - name: date
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/department-productivity:
    get:
      operationId: getDepartmentProductivity
      tags: [dashboard]
      summary: Get department productivity breakdown
      parameters:
        - name: date
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Department productivity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DepartmentProductivity"
...
components:
  schemas:
...
    ReportComment:
      type: object
      properties:
        id: { type: integer }
        reportId: { type: integer }
        userId: { type: integer }
        userName: { type: string }
        userRole: { type: string }
        comment: { type: string }
        createdAt: { type: string }
      required: [id, reportId, userId, userName, userRole, comment, createdAt]
    CommentInput:
      type: object
      properties:
        comment: { type: string }
      required: [comment]
...
    Notification:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        title: { type: string }
        message: { type: string }
        isRead: { type: boolean }
        type: { type: string }
        relatedReportId: { type: ["integer", "null"] }
        createdAt: { type: string }
      required: [id, userId, title, message, isRead, type, createdAt]
    DashboardSummary:
      type: object
      properties:
        totalEmployees: { type: integer }
        submittedToday: { type: integer }
        notSubmittedToday: { type: integer }
        totalTasksToday: { type: integer }
        tasksCompleted: { type: integer }
        tasksPending: { type: integer }
        submitRate: { type: integer }
        completionRate: { type: integer }
      required: [totalEmployees, submittedToday, notSubmittedToday, totalTasksToday, tasksCompleted, tasksPending, submitRate, completionRate]
    DepartmentProductivity:
      type: object
      properties:
        departmentId: { type: integer }
        departmentName: { type: string }
        employeeCount: { type: integer }
        submittedCount: { type: integer }
        avgProgress: { type: integer }
        submitRate: { type: integer }
      required: [departmentId, departmentName, employeeCount, submittedCount, avgProgress, submitRate]
...