# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Laporan Harian - Daily Work Report System API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalEmployees: number;
  submittedToday: number;
  notSubmittedToday: number;
  totalTasksToday: number;
  tasksCompleted: number;
  tasksPending: number;
  submitRate: number;
  completionRate: number;
}