# lib/api-zod/src/generated/types/reportComment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Laporan Harian - Daily Work Report System API
 * OpenAPI spec version: 0.1.0
 */

export interface ReportComment {
  id: number;
  reportId: number;
  userId: number;
  userName: string;
  userRole: string;
  comment: string;
  createdAt: string;
}